modularized lambda handlers
created a folder, lambda handlers, and separated those into their own files
