+++ lambda_handlers/__init__.py
@@ -1,0 +1,1 @@
+"""Lambda handlers package for ACME Package Registry."""
